+++ drivers/gpu/nvgpu/os/posix/posix-nvgpu_mem.c
@@ -114,7 +114,7 @@ struct nvgpu_sgt *nvgpu_sgt_create_from_mem(struct gk20a *g,
 
 int nvgpu_mem_create_from_mem(struct gk20a *g,
 			      struct nvgpu_mem *dest, struct nvgpu_mem *src,
-			      int start_page, int nr_pages)
+			      u64 start_page, int nr_pages)
 {
 	u64 start = start_page * PAGE_SIZE;
 	u64 size = nr_pages * PAGE_SIZE;
